test fixtures for CLI and API
